+++ FROST-Server.HTTP.Common/src/main/java/de/fraunhofer/iosb/ilt/frostserver/http/common/DatabaseStatus.java
@@ -17,10 +17,12 @@
  */
 package de.fraunhofer.iosb.ilt.frostserver.http.common;
 
+import de.fraunhofer.iosb.ilt.frostserver.persistence.PersistenceManager;
 import de.fraunhofer.iosb.ilt.frostserver.persistence.PersistenceManagerFactory;
 import de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings;
 import static de.fraunhofer.iosb.ilt.frostserver.settings.CoreSettings.TAG_CORE_SETTINGS;
 import de.fraunhofer.iosb.ilt.frostserver.util.LiquibaseUser;
+import de.fraunhofer.iosb.ilt.frostserver.util.LiquibaseUtils;
 import de.fraunhofer.iosb.ilt.frostserver.util.exception.UpgradeFailedException;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -71,21 +73,34 @@ public class DatabaseStatus extends HttpServlet {
             out.println("</form></p>");
             out.println("<p><a href='.'>Back...</a></p>");
 
-            for (LiquibaseUser user : coreSettings.getLiquibaseUsers()) {
-                out.print("<h2>");
-                out.print(user.getClass().getName());
-                out.println("</h2>");
-                out.println("<textarea rows=\"10\" style=\"width:95%;\">");
-                String log = checkForUpgrades(user);
-                out.println(log);
-                out.println("</textarea>");
+            PersistenceManager pm = PersistenceManagerFactory.getInstance(coreSettings).create();
+            if (pm instanceof LiquibaseUser) {
+                checkForUpgrades(out, (LiquibaseUser) pm);
             }
+            for (LiquibaseUser user : coreSettings.getLiquibaseUsers()) {
+                checkForUpgrades(out, user);
+            }
+
             out.println("<p>Done. Click the button to execute the listed updates.</p>");
             out.println("</body>");
             out.println("</html>");
         } catch (IOException exc) {
             LOGGER.error("Error writing output to client", exc);
         }
+    }
+
+    public void checkForUpgrades(final PrintWriter out, LiquibaseUser user) {
+        out.print("<h2>");
+        out.print(user.getClass().getName());
+        out.println("</h2>");
+        out.println("<textarea rows=\"10\" style=\"width:95%;\">");
+        String log = checkForUpgrades(user);
+        out.println(log);
+        out.println("</textarea>");
+    }
+
+    private String checkForUpgrades(final LiquibaseUser user) {
+        return user.checkForUpgrades();
     }
 
     protected void processPostRequest(HttpServletRequest request, HttpServletResponse response) {
@@ -102,13 +117,12 @@ public class DatabaseStatus extends HttpServlet {
             out.println("<body>");
             out.println("<h1>Servlet DatabaseStatus at " + request.getContextPath() + "</h1><p>Updating Database</p>");
 
+            PersistenceManager pm = PersistenceManagerFactory.getInstance(coreSettings).create();
+            if (pm instanceof LiquibaseUser) {
+                processUpgrade(out, (LiquibaseUser) pm);
+            }
             for (LiquibaseUser user : coreSettings.getLiquibaseUsers()) {
-                out.print("<h2>");
-                out.print(user.getClass().getName());
-                out.println("</h2>");
-                out.println("<textarea rows=\"10\" style=\"width:95%;\">");
-                processUpgrade(user, out);
-                out.println("</textarea>");
+                processUpgrade(out, user);
             }
 
             out.println("<p>Done. <a href='DatabaseStatus'>Back...</a></p>");
@@ -119,8 +133,13 @@ public class DatabaseStatus extends HttpServlet {
         }
     }
 
-    private String checkForUpgrades(final LiquibaseUser user) {
-        return user.checkForUpgrades();
+    public void processUpgrade(final PrintWriter out, LiquibaseUser user) throws IOException {
+        out.print("<h2>");
+        out.print(user.getClass().getName());
+        out.println("</h2>");
+        out.println("<textarea rows=\"10\" style=\"width:95%;\">");
+        processUpgrade(user, out);
+        out.println("</textarea>");
     }
 
     private void processUpgrade(final LiquibaseUser user, final PrintWriter out) throws IOException {